moved `TimelineSettings` to `crate::timeline::settings`
