+++ phichain-editor/src/editing/clipboard/mod.rs
@@ -5,7 +5,8 @@ use crate::editing::command::{CommandSequence, EditorCommand};
 use crate::editing::DoCommandEvent;
 use crate::hotkey::HotkeyRegistrationExt;
 use crate::selection::{Selected, SelectedLine};
-use crate::tab::timeline::{TimelineSettings, TimelineViewport};
+use crate::tab::timeline::TimelineViewport;
+use crate::timeline::settings::TimelineSettings;
 use crate::timeline::TimelineContext;
 use crate::utils::compat::ControlKeyExt;
 use bevy::prelude::*;
