+++ phichain-editor/src/editing/move_note.rs
@@ -2,7 +2,7 @@ use crate::editing::command::note::EditNote;
 use crate::editing::command::{CommandSequence, EditorCommand};
 use crate::editing::DoCommandEvent;
 use crate::selection::Selected;
-use crate::tab::timeline::TimelineSettings;
+use crate::timeline::settings::TimelineSettings;
 use bevy::prelude::*;
 use num::{FromPrimitive, Rational32};
 use phichain_chart::note::Note;
